update dependencies and refactor control logic
