+++ routes/starttime.ts
@@ -18,6 +18,9 @@ import { parseTime } from '../services/time.ts';
 import { EventLoader } from '../types/index.ts';
 import categories from './categories.ts';
 
+// Regex to validate time format (e.g., 12:25 or 1:25:13)
+const TIME_FORMAT_REGEX = /^\d{1,2}:\d{1,2}(:\d{2})?$/;
+
 interface Category {
   name: string;
   runners: DataPoint[];
@@ -44,7 +47,9 @@ export default function(loader: EventLoader) {
 
         let last: number | null = null;
         let pos = 1;
-        const filtered = category.runners.filter((runner) => parseTime(runner.time) !== null);
+        const filtered = category.runners.filter((runner) => 
+          runner.time && TIME_FORMAT_REGEX.test(runner.time)
+        );
         filtered.forEach((runner, idx) => {
           if (last != null) {
             if ((parseTime(runner.time) ?? 0) > last) {
